chatbot-oryxhr/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pyspellchecker==0.8.2
